move planner tests;
